Massive Update (Read More): This commit marks an important step in the progress of our website, beginning to add some pizazz to the actual history.
Currently, our timeline has events that also function the same. This is good for consistency, but as we begin to work more on this, we will want add some unique visuals/UI/etc in order to better visualize/present the History of Computing.

Thus. we will begin to partially transition off of Event.JS, and use variations of Event.JS like AppleEvent.JS, under the EventTypes component directory. For now the event.js file will remain, so that we can work **both** on the ‘default look’ as well as begin to work on specific styling for specific events.

**TO BE CLEAR: THIS IS A WORK IN PROGRESS AND WE WILL PROBABLY REFACTOR THIS IN THE FUTURE FOR EFFICIENCY AND QUALITY**

To that end: I have begun some of our work:

	1. I have added a fonts.css file, this is a global file (imported into app.css).
		* This allows us to begin to better present specific events.
		* For example, an event about Apple will use the San Fransisco Pro font that they developed, whereas an event about IBM will use the IBM Plex Serif font that they developed.
	2. I have added two types of EventTypes
		1. AppleEvent
			* There is a new animation, gradient, event marker, and font (San Fransisco Pro) for each event for Apple related history.
		2.  IBMEvent
			* Mostly WIP
			* New font (IBM Plex Serif)